adding homework2 information
