+++ homeworks/homework1/.gitignore
@@ -4,3 +4,4 @@
 *.log
 *.fdb_latexmk
 *.synctex.gz
+*.pdf
