+++ homeworks/homework1/CODE_ASSIGNMENT/code_masters/cifar10/get_datasets.py
@@ -1,0 +1,12 @@
+#!/usr/bin/env python3
+
+import urllib.request
+import tarfile
+
+
+url = "https://www.cs.toronto.edu/~kriz/cifar-10-python.tar.gz"
+fileobj = urllib.request.urlopen(url)
+
+print('Downloading...')
+with tarfile.open(fileobj=fileobj, mode="r|gz") as tar:
+    tar.extractall()
