+++ homeworks/homework1/CODE_ASSIGNMENT/code_masters/cifar10/run_for_dataset.sh
@@ -1,0 +1,3 @@
+## running get_datasets.py 
+
+python get_datasets.py
