--- homeworks/homework1/code_masters/cifar10/run_for_dataset.sh
@@ -1,3 +1,0 @@
-## running get_datasets.py 
-
-python get_datasets.py
