+++ homeworks/homework2/homework02_template.aux
@@ -1,0 +1,2 @@
+\relax 
+\gdef \@abspage@last{2}
