+++ homeworks/homework2/homework02_template.fdb_latexmk
@@ -1,0 +1,56 @@
+# Fdb version 3
+["xdvipdfmx"] 1644552691 "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template.xdv" "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template.pdf" "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template" 1644552691
+  "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template.xdv" 1644552691 20872 fe1bf78dd926b25d435c869542703253 "xelatex"
+  (generated)
+  "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template.pdf"
+["xelatex"] 1644552691 "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template.tex" "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template.xdv" "homework02_template" 1644552691
+  "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template.aux" 1644552691 32 044b7f8fc9779af7531264e0c5c84b6d ""
+  "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template.tex" 1644552690 4340 3d72dd48a79dccdceb0b7330165e086b ""
+  "/usr/local/texlive/2021/texmf-dist/fonts/map/fontname/texfonts.map" 1577235249 3524 cb3e574dea2d1052e39280babc910dc8 ""
+  "/usr/local/texlive/2021/texmf-dist/fonts/tfm/public/amsfonts/cmextra/cmex7.tfm" 1246382020 1004 54797486969f23fa377b128694d548df ""
+  "/usr/local/texlive/2021/texmf-dist/fonts/tfm/public/amsfonts/symbols/msam10.tfm" 1246382020 916 f87d7c45f9c908e672703b83b72241a3 ""
+  "/usr/local/texlive/2021/texmf-dist/fonts/tfm/public/amsfonts/symbols/msam7.tfm" 1246382020 928 2dc8d444221b7a635bb58038579b861a ""
+  "/usr/local/texlive/2021/texmf-dist/fonts/tfm/public/amsfonts/symbols/msbm10.tfm" 1246382020 908 2921f8a10601f252058503cc6570e581 ""
+  "/usr/local/texlive/2021/texmf-dist/fonts/tfm/public/amsfonts/symbols/msbm7.tfm" 1246382020 940 228d6584342e91276bf566bcf9716b83 ""
+  "/usr/local/texlive/2021/texmf-dist/fonts/tfm/public/cm/cmex10.tfm" 1136768653 992 662f679a0b3d2d53c1b94050fdaa3f50 ""
+  "/usr/local/texlive/2021/texmf-dist/fonts/tfm/public/cm/cmmi12.tfm" 1136768653 1524 4414a8315f39513458b80dfc63bff03a ""
+  "/usr/local/texlive/2021/texmf-dist/fonts/tfm/public/cm/cmr12.tfm" 1136768653 1288 655e228510b4c2a1abe905c368440826 ""
+  "/usr/local/texlive/2021/texmf-dist/fonts/tfm/public/cm/cmsy10.tfm" 1136768653 1124 6c73e740cf17375f03eec0ee63599741 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/amsfonts/amsfonts.sty" 1359763108 5949 3f3fd50a8cc94c3d4cbf4fc66cd3df1c ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/amsfonts/amssymb.sty" 1359763108 13829 94730e64147574077f8ecfea9bb69af4 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/amsfonts/umsa.fd" 1359763108 961 6518c6525a34feb5e8250ffa91731cff ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/amsfonts/umsb.fd" 1359763108 961 d02606146ba5601b5645f987c92e6193 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/amsmath/amsbsy.sty" 1523134290 2211 ca7ce284ab93c8eecdc6029dc5ccbd73 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/amsmath/amsgen.sty" 1523134290 4161 7f6eb9092061a11f87d08ed13515b48d ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/amsmath/amsmath.sty" 1601675358 87353 2c21ff5f2e32e1bf714e600924d810db ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/amsmath/amsopn.sty" 1523134290 4116 32e6abd27229755a83a8b7f18e583890 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/amsmath/amstext.sty" 1523134290 2432 8ff93b1137020e8f21930562a874ae66 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/base/article.cls" 1601675358 20145 aad8c3dd3bc36e260347b84002182bc2 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/base/size11.clo" 1601675358 8465 2dd6985139dc15bc64ec71e501e76e40 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/base/ts1cmr.fd" 1580683321 2431 b6cead9f069ed55f4fba0c637aa047b6 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/bbm-macros/bbm.sty" 1191314257 1644 1e0d54b051369c3f457872824cac219f ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/doublestroke/dsfont.sty" 1137110062 230 7bc61880b468bfd38aedc173be7c3486 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/float/float.sty" 1137110151 6749 16d2656a1984957e674b149555f1ea1d ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/graphics-cfg/color.cfg" 1459978653 1213 620bba36b25224fa9b7e1ccb4ecb76fd ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/graphics-cfg/graphics.cfg" 1465944070 1224 978390e9c2234eab29404bc21b268d1e ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/graphics-def/xetex.def" 1616189247 15263 06bfc49e20bdfcb827e011063ecb4ae6 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/graphics/graphics.sty" 1601675358 18272 a8c6a275b34ab6717ceeb8fa04b104e2 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/graphics/graphicx.sty" 1601675358 7919 20fdfdd783821971c55bc8ee918cbe63 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/graphics/keyval.sty" 1580683321 2590 e3b24ff953e5b58d924f163d25380312 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/graphics/trig.sty" 1580683321 3976 d7fa7d81d2870d509d25b17d0245e735 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/l3backend/l3backend-xetex.def" 1616101731 34666 58d292eee7df5fb43d9818e3e84b305d ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/listings/listings.cfg" 1585170648 1830 e31effa752c61538383451ae21332364 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/listings/listings.sty" 1585170648 80964 64e57373f36316e4a09b517cbf1aba2e ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/listings/lstmisc.sty" 1585170648 77022 ee25ce086f4a79d8cf73bac6f94c02a5 ""
+  "/usr/local/texlive/2021/texmf-dist/tex/latex/xcolor/xcolor.sty" 1463002160 55589 34128738f682d033422ca125f82e5d62 ""
+  "/usr/local/texlive/2021/texmf-dist/web2c/texmf.cnf" 1616443989 39972 3d6bb4d9d719e93193fa020b58adf5d4 ""
+  "/usr/local/texlive/2021/texmf-var/web2c/xetex/xelatex.fmt" 1630788356 5093946 1133dc7d227a329d9f5c54e07319928b ""
+  "/usr/local/texlive/2021/texmf.cnf" 1616695596 577 4c1911a7228e554c6032256d3cf5567d ""
+  "homework02_template.aux" 1644552691 32 044b7f8fc9779af7531264e0c5c84b6d "xelatex"
+  "homework02_template.tex" 1644552690 4340 3d72dd48a79dccdceb0b7330165e086b ""
+  (generated)
+  "homework02_template.aux"
+  "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template.log"
+  "homework02_template.xdv"
+  "/Users/mrinal18/Desktop/Research/Deep_Learning_Catalyst/homeworks/homework2/homework02_template.xdv"
+  "homework02_template.log"
