+++ homeworks/homework2/homework02_template.tex
@@ -1,0 +1,134 @@
+\documentclass[11pt]{article}
+
+\usepackage{amsmath,amssymb,amsfonts}
+\usepackage{dsfont}
+\usepackage{listings}
+\usepackage{xcolor}
+\usepackage{graphicx}
+\usepackage{bbm}
+\usepackage{float}
+
+
+\definecolor{codegreen}{rgb}{0,0.6,0}
+\definecolor{codegray}{rgb}{0.5,0.5,0.5}
+\definecolor{codepurple}{rgb}{0.58,0,0.82}
+\definecolor{backcolour}{rgb}{0.95,0.95,0.92}
+\lstdefinestyle{mystyle}{
+    backgroundcolor=\color{backcolour},
+    commentstyle=\color{codegreen},
+    keywordstyle=\color{magenta},
+    numberstyle=\tiny\color{codegray},
+    stringstyle=\color{codepurple},
+    basicstyle=\ttfamily\footnotesize,
+    breakatwhitespace=false,
+    breaklines=true,
+    captionpos=b,
+    keepspaces=true,
+    numbers=left,
+    numbersep=5pt,
+    showspaces=false,
+    showstringspaces=false,
+    showtabs=false,
+    tabsize=2
+}
+
+\setlength{\topmargin}{-.5in} \setlength{\textheight}{9.25in}
+\setlength{\oddsidemargin}{0in} \setlength{\textwidth}{6.8in}
+
+%\newcommand*{\SOLVE}{}%
+
+\renewcommand{\vec}[1]{\mbox{\boldmath$#1$}}
+\newcommand{\mm}[1]{\mathbf{#1}}
+
+\newcounter{ProblemNum}
+\newcounter{SubProblemNum}[ProblemNum]
+
+\renewcommand{\theProblemNum}{\arabic{ProblemNum}}
+\renewcommand{\theSubProblemNum}{\alph{SubProblemNum}}
+
+\newcommand*{\anyproblem}[1]{\section*{#1}}
+\newcommand*{\problem}[1]{\stepcounter{ProblemNum} %
+   \anyproblem{Problem \theProblemNum \; (#1 points)}}
+\newcommand*{\soln}[1]{\subsection*{#1}}
+\newcommand*{\solution}{\soln{Solution}}
+\newenvironment{solutions}
+  {\section[Solution]{\textcolor{red}{Solution}}\color{red}}
+  {\normalcolor}
+\renewcommand*{\part}{\stepcounter{SubProblemNum} %
+  \soln{Part (\theSubProblemNum)}}
+\renewcommand{\theenumi}{(\alph{enumi})}
+\renewcommand{\labelenumi}{\theenumi}
+\renewcommand{\theenumii}{\roman{enumii}}
+\let\endsection\relax
+\let\endsubsection\relax
+
+\graphicspath{
+{.}
+}
+\lstset{style=mystyle}
+
+\begin{document}
+
+\Large
+\noindent{\bf CS4851/6851 IDL: Homework 1 \hfill \today}
+\medskip\hrule
+
+\vspace{20pt}
+
+Note: All coding problems are to be implemented in a single jupyter notebook.
+
+Note: this is the distribution of questions:
+\begin{enumerate}
+  \item Question 1 to Question6: Required for everyone.
+  \item Question 7 - Question 8: Required only for Graduate Students
+  \item Question 9- Question 10: Bonus marks
+\end{enumerate}
+
+\vspace{20pt}
+\noindent\rule[0.5ex]{0.45\linewidth}{1pt} Bonus for undergraduates
+beyond this line
+\problem{5} We studies Automatic Differentiation in class. Consider the following function:
+
+\begin{equation}
+  f(x, y, z)=   \frac{1}{3} (x_1x_2\sin(x_3) + exp^{x_1x_2})
+\end{equation}
+
+\begin{enumerate}
+  \item Draw the computation graph for this equation.
+  \item List the detailed computation steps for forward and backward mode of AD. Provide your answers the same way we did in class, by using notations like: 
+      \begin{equation}
+        v_{-2}, v_{-1}, v_{0}, v_{1}, v_{2}
+      \end{equation}
+
+  to get you started:
+  \begin{equation}
+    v_{-2} = 1, \dot{v}_{-2} = \frac{\partial{v_{-2}}}{\partial{x_{-2}}} = 0
+  \end{equation}
+
+  Provide all the steps like this and values for other nodes in computation graph.
+\end{enumerate}
+
+\problem{5} Select the correct option.Accuracy is probably the most commonly used metric in the most common type of Deep 
+learning problems:
+
+\begin{enumerate}
+  \item Accuracy is very easy to interpret and is aligned on what you want to measure.
+  \item Accuracy is differential and can be used to directly in gradient based optimisation process
+  \item We want accuracy to count every error as equal.
+  \item If the number of sample is similar, we cannot make any mistake on the less populated category.
+\end{enumerate}
+
+
+\problem{5} You are working as a Machine Learning Engineer in Metflix Inc. You are building a model to classify users who watch a lot of movies in Ultiverse. What metrics will you choose to evaluate your model?
+
+\problem{5} You are working as a Machine Learning Engineer in Metflix Inc. You are building a model to classify users who watch a lot of movies in Ultiverse. What metrics will you choose to evaluate your model?
+
+\problem{5} Which method is used involved in numerical optimization of an appropriate selection of model criterion? How do you define the error of such estimator? 
+
+\end{document}
+
+%%% Local Variables:
+%%% mode: latex
+%%% TeX-master: t
+%%% End:
+
